Move _feed_data_to_bufferred_proto() to protocols
